Fastq files are now moved to buffer, properties file drafted
Some global variables were defined, so only current sample line
is passed to process_sample.

Other global variables that must probably be defined are

- Investigator Name
- Experiment Name
- ...

i.e. all those that are defined in the sections before samples.

`data.properties` file is written but needs to be completed

The backup section still needs to be written.